+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * New cop `LeadingCommentSpace` checks for missing space after `#` in comments.
 * New cop `ColonMethodCall` tracks uses of `::` for method calls.
 * New cop `AvoidGlobalVars` tracks uses of non built-in global variables.
+* New cop `SpaceAfterControlKeyword` tracks missing spaces after `if/elsif/case/when/until/unless/while`.
 
 ### Bugs fixed